requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
absl-py==2.2.2
    # via tensorboard
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via tts
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyascii==0.3.2
    # via tts
async-timeout==5.0.1
    # via aiohttp
attrs==25.3.0
    # via aiohttp
audioread==3.0.1
    # via librosa
babel==2.17.0
    # via gruut
bangla==0.0.5
    # via tts
blinker==1.9.0
    # via flask
blis==1.2.1
    # via thinc
bnnumerizer==0.0.2
    # via tts
bnunicodenormalizer==0.1.7
    # via tts
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2025.4.26
    # via requests
cffi==1.17.1
    # via soundfile
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via
    #   flask
    #   nltk
    #   typer
cloudpathlib==0.21.1
    # via weasel
confection==0.1.5
    # via
    #   thinc
    #   weasel
contourpy==1.2.1
    # via matplotlib
coqpit==0.0.17
    # via
    #   trainer
    #   tts
cycler==0.12.1
    # via matplotlib
cymem==2.0.11
    # via
    #   preshed
    #   spacy
    #   thinc
cython==3.1.0
    # via tts
dateparser==1.1.8
    # via gruut
decorator==5.2.1
    # via librosa
docopt==0.6.2
    # via num2words
einops==0.8.1
    # via
    #   encodec
    #   tts
encodec==0.1.1
    # via tts
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flask==3.1.1
    # via
    #   -r requirements.in
    #   tts
fonttools==4.58.0
    # via matplotlib
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.3.2
    # via
    #   huggingface-hub
    #   torch
    #   trainer
    #   tts
g2pkk==0.1.2
    # via tts
grpcio==1.71.0
    # via tensorboard
gruut[de,es,fr]==2.2.3
    # via tts
gruut-ipa==0.13.0
    # via gruut
gruut-lang-de==2.0.1
    # via gruut
gruut-lang-en==2.0.1
    # via gruut
gruut-lang-es==2.0.1
    # via gruut
gruut-lang-fr==2.0.2
    # via gruut
hangul-romanize==0.1.0
    # via tts
huggingface-hub==0.31.2
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   requests
    #   yarl
inflect==7.5.0
    # via tts
itsdangerous==2.2.0
    # via flask
jamo==0.4.1
    # via
    #   g2pkk
    #   tts
jieba==0.42.1
    # via tts
jinja2==3.1.6
    # via
    #   flask
    #   spacy
    #   torch
joblib==1.5.0
    # via
    #   librosa
    #   nltk
    #   pynndescent
    #   scikit-learn
jsonlines==1.2.0
    # via gruut
kiwisolver==1.4.8
    # via matplotlib
langcodes==3.5.0
    # via spacy
language-data==1.3.0
    # via langcodes
lazy-loader==0.4
    # via librosa
librosa==0.10.0
    # via tts
llvmlite==0.43.0
    # via
    #   numba
    #   pynndescent
marisa-trie==1.2.1
    # via language-data
markdown==3.8
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
matplotlib==3.8.4
    # via tts
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.7.0
    # via
    #   inflect
    #   openai-whisper
mpmath==1.3.0
    # via sympy
msgpack==1.1.0
    # via librosa
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
murmurhash==1.0.12
    # via
    #   preshed
    #   spacy
    #   thinc
networkx==2.8.8
    # via
    #   gruut
    #   torch
nltk==3.9.1
    # via
    #   g2pkk
    #   tts
num2words==0.5.14
    # via
    #   gruut
    #   tts
numba==0.60.0
    # via
    #   librosa
    #   openai-whisper
    #   pynndescent
    #   tts
    #   umap-learn
numpy==1.22.0
    # via
    #   blis
    #   contourpy
    #   encodec
    #   gruut
    #   librosa
    #   matplotlib
    #   numba
    #   openai-whisper
    #   pandas
    #   scikit-learn
    #   scipy
    #   soundfile
    #   soxr
    #   spacy
    #   tensorboard
    #   thinc
    #   transformers
    #   tts
    #   umap-learn
openai-whisper==20240930
    # via -r requirements.in
packaging==25.0
    # via
    #   huggingface-hub
    #   lazy-loader
    #   matplotlib
    #   pooch
    #   spacy
    #   tensorboard
    #   thinc
    #   transformers
    #   tts
    #   weasel
pandas==1.5.3
    # via tts
pillow==11.2.1
    # via matplotlib
platformdirs==4.3.8
    # via pooch
pooch==1.8.2
    # via librosa
preshed==3.0.9
    # via
    #   spacy
    #   thinc
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.0
    # via tensorboard
psutil==7.0.0
    # via trainer
pycparser==2.22
    # via cffi
pydantic==2.11.4
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via rich
pynndescent==0.5.13
    # via umap-learn
pyparsing==3.2.3
    # via matplotlib
pypinyin==0.54.0
    # via tts
pysbd==0.3.4
    # via tts
python-crfsuite==0.9.11
    # via gruut
python-dateutil==2.9.0.post0
    # via
    #   dateparser
    #   matplotlib
    #   pandas
pytz==2025.2
    # via
    #   dateparser
    #   pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
    #   tts
regex==2024.11.6
    # via
    #   dateparser
    #   nltk
    #   tiktoken
    #   transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   pooch
    #   spacy
    #   tiktoken
    #   transformers
    #   weasel
rich==14.0.0
    # via typer
safetensors==0.5.3
    # via transformers
scikit-learn==1.6.1
    # via
    #   librosa
    #   pynndescent
    #   tts
    #   umap-learn
scipy==1.11.4
    # via
    #   librosa
    #   pynndescent
    #   scikit-learn
    #   tts
    #   umap-learn
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   jsonlines
    #   python-dateutil
    #   tensorboard
smart-open==7.1.0
    # via weasel
soundfile==0.13.1
    # via
    #   librosa
    #   trainer
    #   tts
soxr==0.5.0.post1
    # via librosa
spacy[ja]==3.8.5
    # via tts
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.5.1
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
sudachidict-core==20250129
    # via spacy
sudachipy==0.6.10
    # via
    #   spacy
    #   sudachidict-core
sympy==1.14.0
    # via torch
tensorboard==2.19.0
    # via trainer
tensorboard-data-server==0.7.2
    # via tensorboard
thinc==8.3.4
    # via spacy
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.9.0
    # via openai-whisper
tokenizers==0.21.1
    # via transformers
torch==2.2.2
    # via
    #   -r requirements.in
    #   encodec
    #   openai-whisper
    #   torchaudio
    #   trainer
    #   tts
torchaudio==2.2.2
    # via
    #   encodec
    #   tts
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   nltk
    #   openai-whisper
    #   spacy
    #   transformers
    #   tts
    #   umap-learn
trainer==0.0.36
    # via tts
transformers==4.51.3
    # via
    #   -r requirements.in
    #   tts
tts==0.22.0
    # via -r requirements.in
typeguard==4.4.2
    # via inflect
typer==0.15.4
    # via
    #   spacy
    #   weasel
typing-extensions==4.13.2
    # via
    #   cloudpathlib
    #   huggingface-hub
    #   librosa
    #   multidict
    #   pydantic
    #   pydantic-core
    #   rich
    #   torch
    #   typeguard
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzlocal==5.3.1
    # via dateparser
umap-learn==0.5.7
    # via tts
unidecode==1.4.0
    # via tts
urllib3==2.4.0
    # via requests
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.1
    # via spacy
werkzeug==3.1.3
    # via
    #   flask
    #   tensorboard
wrapt==1.17.2
    # via smart-open
yarl==1.20.0
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
